fx: seperate HoverHighlight
